+++ Tests/Runtime/HttpClient/PostRequests.cs
@@ -1,9 +1,13 @@
 using System;
 using System.Collections;
+using System.Collections.Generic;
 using System.Net;
+using KindMen.Uxios.Http;
 using Newtonsoft.Json;
 using NUnit.Framework;
+using UnityEngine;
 using UnityEngine.TestTools;
+using Random = System.Random;
 
 namespace KindMen.Uxios.Tests.HttpClient
 {
@@ -21,10 +25,48 @@ namespace KindMen.Uxios.Tests.HttpClient
             public string body;
         }
 
+        private class HttpBinPostResponse
+        {
+            public string data;
+        }
+
         [SetUp]
         public void SetUp()
         {
             uxios = new Uxios();
+        }
+
+        [UnityTest]
+        public IEnumerator PostsRawData()
+        {
+            var url = new Uri("https://httpbin.org/post");
+
+            var data = new byte[1024];
+            Random random = new Random();
+            random.NextBytes(data);
+
+            var promise = uxios.Post<byte[], HttpBinPostResponse>(url, data);
+
+            yield return PromiseAssertions.AssertPromiseSucceeds(
+                promise, 
+                response =>
+                {
+                    Assert.That(response.Request.Headers, Contains.Key(Headers.ContentType));
+                    Assert.That(response.Request.Headers[Headers.ContentType], Is.EqualTo("application/octet-stream"));
+
+                    Assert.That(response.Status, Is.EqualTo(HttpStatusCode.OK));
+                    Assert.That(response.Data, Is.TypeOf<HttpBinPostResponse>());
+                    
+                    HttpBinPostResponse post = response.Data as HttpBinPostResponse;
+                
+                    var expectedData = "data:application/octet-stream;base64," + Convert.ToBase64String(data);
+    
+                    Assert.That(post, Is.Not.Null);
+                    Assert.That(post.data, Is.EqualTo(expectedData));
+                    Assert.That(response.Headers, Contains.Key(Headers.ContentType));
+                    Assert.That(response.Headers[Headers.ContentType], Is.EqualTo("application/json"));
+                }
+            );
         }
 
         [UnityTest]
@@ -47,6 +89,8 @@ namespace KindMen.Uxios.Tests.HttpClient
                 {
                     Assert.That(response.Status, Is.EqualTo(HttpStatusCode.Created));
                     Assert.That(response.Data, Is.TypeOf<ExamplePost>());
+                    Assert.That(response.Request.Headers, Contains.Key(Headers.ContentType));
+                    Assert.That(response.Request.Headers[Headers.ContentType], Is.EqualTo("application/json"));
                     
                     ExamplePost post = response.Data as ExamplePost;
                 
@@ -55,8 +99,8 @@ namespace KindMen.Uxios.Tests.HttpClient
                     Assert.That(post.userId, Is.EqualTo(data.userId));
                     Assert.That(post.title, Is.EqualTo(data.title));
                     Assert.That(post.body, Is.EqualTo(data.body));
-                    Assert.That(response.Headers, Contains.Key("Content-Type"));
-                    Assert.That(response.Headers["Content-Type"], Is.EqualTo("application/json; charset=utf-8"));
+                    Assert.That(response.Headers, Contains.Key(Headers.ContentType));
+                    Assert.That(response.Headers[Headers.ContentType], Is.EqualTo("application/json; charset=utf-8"));
                 }
             );
         }